# docs/api-docs.yaml
openapi: 3.0.0
info:
    title: FileSharing API
    description: API for uploading and managing files.
    version: 1.0.0
servers:
    - url: http://localhost:3000/api
      description: Development server
    - url: https://your-domain.com/api
      description: Production server
...
components:
    securitySchemes:
        bearerAuth:
            type: http
            scheme: bearer
            bearerFormat: JWT
            description: "JWT authorization token obtained during file upload"
    schemas:
...
        Error:
            type: object
            properties:
                status:
                    type: string
                    enum: [error]
                code:
                    type: integer
                    description: HTTP status code
                message:
                    type: string
                    description: Error message
                errors:
                    type: array
                    items:
                        type: string
                    description: Additional error details
            example:
                status: "error"
                code: 400
                message: "Invalid input data"
                errors:
                    [
                        "Filename is required",
                        "Maximum downloads must be a positive number",
                    ]
...
security:
    - bearerAuth: []
paths:
...
    /files/{id}:
        get:
            summary: Get upload by ID
            description: Retrieve details about a specific upload by its ID
            tags:
                - Files
            parameters:
                - in: path
                  name: id
                  required: true
                  schema:
                      type: string
                  description: Upload ID
            responses:
                "200":
                    description: Upload details
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    status:
                                        type: string
                                        enum: [success]
                                    code:
                                        type: integer
                                    data:
                                        type: object
                                        properties:
                                            id:
                                                type: string
                                            originalName:
                                                type: string
                                            customName:
                                                type: string
                                            mimeType:
                                                type: string
                                            size:
                                                type: integer
                                            hasPassword:
                                                type: boolean
                                            maxDownloads:
                                                type: integer
                                            downloadsCount:
                                                type: integer
                                            expiresAt:
                                                type: string
                                                format: date-time
                                            isExpired:
                                                type: boolean
                                                description: Expiration status flag
                                            createdAt:
                                                type: string
                                                format: date-time
                                            updatedAt:
                                                type: string
                                                format: date-time
                "404":
                    description: Upload not found
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Error"
                "429":
                    description: Too many requests
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Error"
                "500":
                    description: Server error
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Error"

        patch:
            summary: Update file metadata
            security:
                - bearerAuth: []
            description: |
                Update file metadata.
                Requires a JWT token (delete_token) in the Authorization header: Bearer <token>.
                Available only within 24 hours after upload.
            tags:
                - Files
            parameters:
                - in: path
                  name: id
                  required: true
                  schema:
                      type: string
                  description: Upload ID
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            properties:
                                filename:
                                    type: string
                                    description: New filename
                                password:
                                    type: string
                                    description: New password (or remove if null)
                                maxDownloads:
                                    type: integer
                                    description: New maximum number of downloads
                                expiresIn:
                                    type: integer
                                    minimum: 0
                                    example: 24
                                    description: New expiration time in hours (0 to remove restriction)
                        examples:
                            setExpiration:
                                summary: Set expiration time
                                value:
                                    expiresIn: 24
                            removeExpiration:
                                summary: Remove expiration time
                                value:
                                    expiresIn: 0
            responses:
                "200":
                    description: File updated successfully
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    status:
                                        type: string
                                        enum: [success]
                                    code:
                                        type: integer
                                    data:
                                        type: object
                                        properties:
                                            message:
                                                type: string
                                            id:
                                                type: string
                                            filename:
                                                type: string
                                            updated:
                                                type: boolean
                "401":
                    description: Unauthorized, missing or invalid token
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Error"
                "403":
                    description: Forbidden, 24-hour modification period expired
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Error"
                "404":
                    description: Upload not found
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Error"
                "429":
                    description: Too many requests
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Error"
                "500":
                    description: Server error
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Error"

        delete:
            summary: Delete file
            security:
                - bearerAuth: []
            description: |
                Delete the file and all its metadata.
                Requires a JWT token (delete_token) in the Authorization header: Bearer <token>.
                Available only within 24 hours after upload.
            tags:
                - Files
            parameters:
                - in: path
                  name: id
                  required: true
                  schema:
                      type: string
                  description: Upload ID
            responses:
                "200":
                    description: File deleted successfully
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    status:
                                        type: string
                                        enum: [success]
                                    code:
                                        type: integer
                                    data:
                                        type: object
                                        properties:
                                            message:
                                                type: string
                                            id:
                                                type: string
                                            deleted:
                                                type: boolean
                "401":
                    description: Unauthorized, missing or invalid token
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Error"
                "403":
                    description: Forbidden, 24-hour deletion period expired
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Error"
                "404":
                    description: Upload not found
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Error"
                "429":
                    description: Too many requests
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Error"
                "500":
                    description: Server error
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Error"
    /files/{id}/download:
        get:
            summary: Download file
            description: |
                Download a file by ID.
                If the file is password-protected, the password parameter must be provided.
                Does not require authorization.
            tags:
                - Files
            parameters:
                - in: path
                  name: id
                  required: true
                  schema:
                      type: string
                  description: Upload ID
                - in: query
                  name: password
                  required: false
                  schema:
                      type: string
                  description: Password for protected files
            responses:
                "200":
                    description: File stream
                    content:
                        application/octet-stream:
                            schema:
                                type: string
                                format: binary
                "401":
                    description: Password required or invalid password
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Error"
                "403":
                    description: Download limit exceeded
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Error"
                "404":
                    description: File not found
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Error"
                "410":
                    description: File expired
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Error"
                "429":
                    description: Too many requests
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Error"
                "500":
                    description: Server error
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Error"